Refactor main module
