+++ src/main.rs
@@ -1,19 +1,12 @@
 #![forbid(unsafe_code)]
 #![deny(clippy::all, clippy::pedantic, clippy::nursery, clippy::cargo)]
 
-use std::ffi::OsStr;
-use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::process::ExitCode;
 
 use clap::Parser;
-use ignore::WalkBuilder;
 use rayon::prelude::*;
-
-struct NullSink;
-impl<'i> saphyr_parser::EventReceiver<'i> for NullSink {
-    fn on_event(&mut self, _ev: saphyr_parser::Event<'i>) {}
-}
+use ryl::{gather_yaml_from_dir, parse_yaml_file};
 
 #[derive(Parser, Debug)]
 #[command(name = "ryl", version, about = "Fast YAML linter written in Rust")]
@@ -21,54 +14,6 @@ struct Cli {
     /// One or more paths: files and/or directories
     #[arg(value_name = "PATH_OR_FILE", num_args = 1..)]
     inputs: Vec<PathBuf>,
-}
-
-fn is_yaml_path(path: &Path) -> bool {
-    matches!(
-        path.extension().and_then(OsStr::to_str).map(str::to_ascii_lowercase),
-        Some(ref ext) if ext == "yml" || ext == "yaml"
-    )
-}
-
-fn gather_yaml_from_dir(dir: &Path) -> Vec<PathBuf> {
-    let mut files = Vec::new();
-    let walker = WalkBuilder::new(dir)
-        .hidden(false)
-        .ignore(true)
-        .git_ignore(true)
-        .git_global(true)
-        .git_exclude(true)
-        .follow_links(false)
-        .build();
-
-    for entry in walker.flatten() {
-        let p = entry.path();
-        if p.is_file() && is_yaml_path(p) {
-            files.push(p.to_path_buf());
-        }
-    }
-    files
-}
-
-fn parse_yaml_file(path: &Path) -> Result<(), String> {
-    let content = fs::read_to_string(path)
-        .map_err(|e| format!("failed to read {}: {}", path.display(), e))?;
-    let mut parser = saphyr_parser::Parser::new_from_str(&content);
-    let mut sink = NullSink;
-    match parser.load(&mut sink, true) {
-        Ok(()) => Ok(()),
-        Err(e) => {
-            let m = e.marker();
-            let msg = e.info();
-            Err(format!(
-                "{}\n  {}:{}       error    syntax error: {} (syntax)",
-                path.display(),
-                m.line(),
-                m.col() + 1,
-                msg
-            ))
-        }
-    }
 }
 
 fn main() -> ExitCode {
